move moderation list screen to new package
