+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/MainActivity.kt
@@ -30,6 +30,7 @@ import com.bounswe.predictionpolls.ui.main.mainScreen
 import com.bounswe.predictionpolls.ui.main.navigateToMainScreen
 import com.bounswe.predictionpolls.ui.profile.myProfileScreen
 import com.bounswe.predictionpolls.ui.moderation.moderationScreen
+import com.bounswe.predictionpolls.ui.moderation.list.moderationScreen
 import com.bounswe.predictionpolls.ui.profile.profileScreen
 import com.bounswe.predictionpolls.ui.signup.signupScreen
 import com.bounswe.predictionpolls.ui.theme.PredictionPollsTheme
