--- prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/ui/moderation/ModerationScreenEvent.kt
@@ -1,3 +1,0 @@
-package com.bounswe.predictionpolls.ui.moderation
-
-sealed class ModerationScreenEvent {}
