+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/ui/moderation/list/ModerationScreenEvent.kt
@@ -1,0 +1,3 @@
+package com.bounswe.predictionpolls.ui.moderation.list
+
+sealed class ModerationScreenEvent {}
