+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/ui/moderation/list/ModerationScreenNavigation.kt
@@ -1,4 +1,4 @@
-package com.bounswe.predictionpolls.ui.moderation
+package com.bounswe.predictionpolls.ui.moderation.list
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
